优化代码

+++ pidautotuner.h
@@ -26,15 +26,15 @@ typedef enum
 typedef struct
 {
   float targetInputValue;
-  unsigned int loopInterval;
+  float loopInterval;
   float minOutput, maxOutput;
   ZNMode znMode;
-  unsigned int cycles;
+  int cycles;
 
-  unsigned int i;
+  int i;
   bool output;
   float outputValue;
-  unsigned int microseconds, t1, t2, tHigh, tLow;
+  int microseconds, t1, t2, tHigh, tLow;
   float max, min;
   float pAverage, iAverage, dAverage;
   float kp, ki, kd;
@@ -68,18 +68,18 @@ typedef struct
 // tuningCycles: number of cycles that the tuning runs for (optional, default is 10)
 extern void vPIDAutotunerInit(PIDAutotuner_t *pxPID);
 extern void vPIDAutotunerSetTargetInputValue(PIDAutotuner_t *pxPID, float target);
-extern void vPIDAutotunerSetLoopInterval(PIDAutotuner_t *pxPID, unsigned int interval);
+extern void vPIDAutotunerSetLoopInterval(PIDAutotuner_t *pxPID, int interval);
 extern void vPIDAutotunerSetOutputRange(PIDAutotuner_t *pxPID, float min, float max);
 extern void vPIDAutotunerSetZNMode(PIDAutotuner_t *pxPID, ZNMode zn);
-extern void vPIDAutotunerSetTuningCycles(PIDAutotuner_t *pxPID, unsigned int tuneCycles);
+extern void vPIDAutotunerSetTuningCycles(PIDAutotuner_t *pxPID, int tuneCycles);
 
 // Must be called immediately before the tuning loop starts
-extern void vPIDAutotunerStartTuningLoop(PIDAutotuner_t *pxPID, long long us);
+extern void vPIDAutotunerStartTuningLoop(PIDAutotuner_t *pxPID, unsigned int us);
 
 // Automatically tune PID
 // This function must be run in a loop at the same speed as the PID loop being tuned
 // See README for more details - https://github.com/jackw01/arduino-pid-autotuner/blob/master/README.md
-extern float fPIDAutotunerTunePID(PIDAutotuner_t *pxPID, float input, long long us);
+extern float fPIDAutotunerTunePID(PIDAutotuner_t *pxPID, float input, unsigned int us);
 
 // Get results of most recent tuning
 extern float fPIDAutotunerGetKp(PIDAutotuner_t *pxPID);
